Formatted Code

+++ contracts/swaps/IDexRecords.sol
@@ -1,6 +1,10 @@
 pragma solidity ^0.5.17;
 
-interface IDexRecords{
-	function retreiveDexAddress(uint dexNumber) external view returns(address);
-	function setDexID(address dexAddress) external;
+interface IDexRecords {
+    function retreiveDexAddress(uint256 dexNumber)
+        external
+        view
+        returns (address);
+
+    function setDexID(address dexAddress) external;
 }
